Extract ActivePlayerTracker, PieceInitializer, and MatchStatusUpdater
