+++ src/main/java/com/shepherdjerred/capstone/logic/board/BoardSettings.java
@@ -1,5 +1,6 @@
 package com.shepherdjerred.capstone.logic.board;
 
+import com.shepherdjerred.capstone.logic.match.MatchSettings.PlayerCount;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
@@ -11,10 +12,12 @@ public final class BoardSettings {
 
   private final int boardSize;
   private final int gridSize;
+  private final PlayerCount playerCount;
 
-  public BoardSettings(int boardSize) {
+  public BoardSettings(int boardSize, PlayerCount playerCount) {
     this.boardSize = boardSize;
     this.gridSize = boardSize * 2 - 1;
+    this.playerCount = playerCount;
 
     if (boardSize % 2 != 1) {
       throw new IllegalArgumentException("Board gridSize must be odd");
