+++ src/main/java/com/shepherdjerred/capstone/logic/match/MatchStatusUpdater.java
@@ -1,0 +1,46 @@
+package com.shepherdjerred.capstone.logic.match;
+
+import com.shepherdjerred.capstone.logic.match.MatchStatus.Status;
+import com.shepherdjerred.capstone.logic.player.Player;
+import com.shepherdjerred.capstone.logic.player.exception.InvalidPlayerException;
+import com.shepherdjerred.capstone.logic.turn.MovePawnTurn;
+import com.shepherdjerred.capstone.logic.turn.Turn;
+
+public class MatchStatusUpdater {
+  public MatchStatus updateMatchStatus(Turn turn, Match match) {
+    var player = turn.getCauser();
+    if (turn instanceof MovePawnTurn) {
+      var board = match.getBoard();
+      var boardSettings = board.getBoardSettings();
+      var gridSize = boardSettings.getGridSize();
+      var movePawnTurn = (MovePawnTurn) turn;
+      var destination = movePawnTurn.getDestination();
+
+      switch (player) {
+        case ONE:
+          if (destination.getY() == gridSize - 1) {
+            return new MatchStatus(Player.ONE, Status.VICTORY);
+          }
+          break;
+        case TWO:
+          if (destination.getY() == 0) {
+            return new MatchStatus(Player.TWO, Status.VICTORY);
+          }
+          break;
+        case THREE:
+          if (destination.getX() == gridSize - 1) {
+            return new MatchStatus(Player.THREE, Status.VICTORY);
+          }
+          break;
+        case FOUR:
+          if (destination.getX() == 0) {
+            return new MatchStatus(Player.FOUR, Status.VICTORY);
+          }
+          break;
+        default:
+          throw new InvalidPlayerException(player);
+      }
+    }
+    return match.getMatchStatus();
+  }
+}
