+++ src/main/java/com/shepherdjerred/capstone/logic/turn/enactor/MatchTurnEnactor.java
@@ -1,18 +1,14 @@
 package com.shepherdjerred.capstone.logic.turn.enactor;
 
+import com.shepherdjerred.capstone.logic.match.ActivePlayerTracker;
 import com.shepherdjerred.capstone.logic.match.Match;
-import com.shepherdjerred.capstone.logic.match.MatchSettings.PlayerCount;
-import com.shepherdjerred.capstone.logic.match.MatchStatus;
-import com.shepherdjerred.capstone.logic.match.MatchStatus.Status;
-import com.shepherdjerred.capstone.logic.player.Player;
-import com.shepherdjerred.capstone.logic.player.exception.InvalidPlayerException;
-import com.shepherdjerred.capstone.logic.turn.MovePawnTurn;
+import com.shepherdjerred.capstone.logic.match.MatchStatusUpdater;
 import com.shepherdjerred.capstone.logic.turn.PlaceWallTurn;
 import com.shepherdjerred.capstone.logic.turn.Turn;
 import com.shepherdjerred.capstone.logic.turn.exception.InvalidTurnException;
 import com.shepherdjerred.capstone.logic.turn.validator.TurnValidator;
 
-// TODO this needs a lot of clean up
+// TODO this needs to be cleaned up
 public class MatchTurnEnactor {
 
   private final TurnEnactorFactory turnEnactorFactory;
@@ -38,9 +34,9 @@ public class MatchTurnEnactor {
     var board = match.getBoard();
 
     var newBoard = enactor.enactTurn(turn, board);
-    var newMatchStatus = updateMatchStatus(turn, match);
+    var newMatchStatus = new MatchStatusUpdater().updateMatchStatus(turn, match);
 
-    var newWallPool = match.getWallPool();
+    var newWallPool = match.getPlayerWallBank();
     if (turn instanceof PlaceWallTurn) {
       newWallPool = newWallPool.takeWall(turn.getCauser());
     }
@@ -50,87 +46,7 @@ public class MatchTurnEnactor {
 
     var matchSettings = match.getMatchSettings();
 
-    return new Match(newBoard, matchSettings, getNextActivePlayer(match),
+    return new Match(newBoard, matchSettings, new ActivePlayerTracker().getNextActivePlayer(match),
         newWallPool, newMatchStatus, newHistory);
-  }
-
-
-  private MatchStatus updateMatchStatus(Turn turn, Match match) {
-    var player = turn.getCauser();
-    if (turn instanceof MovePawnTurn) {
-      var board = match.getBoard();
-      var boardSettings = board.getBoardSettings();
-      var gridSize = boardSettings.getGridSize();
-      var movePawnTurn = (MovePawnTurn) turn;
-      var destination = movePawnTurn.getDestination();
-
-      switch (player) {
-        case ONE:
-          if (destination.getY() == gridSize - 1) {
-            return new MatchStatus(Player.ONE, Status.VICTORY);
-          }
-          break;
-        case TWO:
-          if (destination.getY() == 0) {
-            return new MatchStatus(Player.TWO, Status.VICTORY);
-          }
-          break;
-        case THREE:
-          if (destination.getX() == gridSize - 1) {
-            return new MatchStatus(Player.THREE, Status.VICTORY);
-          }
-          break;
-        case FOUR:
-          if (destination.getX() == 0) {
-            return new MatchStatus(Player.FOUR, Status.VICTORY);
-          }
-          break;
-        default:
-          throw new InvalidPlayerException(player);
-      }
-    }
-    return match.getMatchStatus();
-  }
-
-  /**
-   * Returns the player who comes after the active player
-   */
-  public Player getNextActivePlayer(Match match) {
-    var activePlayer = match.getActivePlayer();
-    var playerCount = match.getMatchSettings().getPlayerCount();
-
-    if (playerCount == PlayerCount.TWO) {
-      return getNextActivePlayerForTwoPlayerMatch(activePlayer);
-    } else if (playerCount == PlayerCount.FOUR) {
-      return getNextActivePlayerForFourPlayerMatch(activePlayer);
-    } else {
-      throw new IllegalStateException("Unknown player count " + playerCount);
-    }
-  }
-
-  private Player getNextActivePlayerForTwoPlayerMatch(Player activePlayer) {
-    switch (activePlayer) {
-      case ONE:
-        return Player.TWO;
-      case TWO:
-        return Player.ONE;
-      default:
-        throw new InvalidPlayerException(activePlayer);
-    }
-  }
-
-  private Player getNextActivePlayerForFourPlayerMatch(Player activePlayer) {
-    switch (activePlayer) {
-      case ONE:
-        return Player.TWO;
-      case TWO:
-        return Player.THREE;
-      case THREE:
-        return Player.FOUR;
-      case FOUR:
-        return Player.ONE;
-      default:
-        throw new InvalidPlayerException(activePlayer);
-    }
   }
 }
