+++ src/test/java/com/shepherdjerred/capstone/logic/MyTest.java
@@ -7,6 +7,7 @@ import com.shepherdjerred.capstone.logic.board.layout.BoardLayout;
 import com.shepherdjerred.capstone.logic.match.Match;
 import com.shepherdjerred.capstone.logic.player.Player;
 import com.shepherdjerred.capstone.logic.turn.MovePawnTurn.MoveType;
+import com.shepherdjerred.capstone.logic.turn.enactor.MatchTurnEnactor;
 import com.shepherdjerred.capstone.logic.turn.validator.TurnValidator;
 import com.shepherdjerred.capstone.logic.util.MatchFormatter;
 import com.shepherdjerred.capstone.logic.match.MatchSettings;
@@ -22,17 +23,16 @@ public class MyTest {
 
   @Test
   public void myTest() {
-    var boardSettings = new BoardSettings(9);
+    var enactor = new MatchTurnEnactor(new TurnEnactorFactory(), new TurnValidator());
+
+    var boardSettings = new BoardSettings(9, PlayerCount.TWO);
     var matchSettings = new MatchSettings(10,
-        PlayerCount.TWO,
-        Player.ONE);
+        Player.ONE,
+        boardSettings);
 
     var initialMatchState = Match.startNewMatch(matchSettings,
         Board.createNewBoard(BoardLayout.fromBoardSettings(boardSettings),
-            boardSettings,
-            matchSettings.getPlayerCount()),
-        new TurnEnactorFactory(),
-        new TurnValidator());
+            boardSettings));
 
     var turn1 = new MovePawnTurn(Player.ONE,
         MoveType.NORMAL,
@@ -50,9 +50,9 @@ public class MyTest {
         new Coordinate(8, 14),
         new Coordinate(8, 12));
 
-    var matchStateAfterTurn1 = initialMatchState.doTurn(turn1);
-    var matchStateAfterTurn2 = matchStateAfterTurn1.doTurn(turn2);
-    var matchStateAfterTurn3 = matchStateAfterTurn2.doTurn(turn3);
+    var matchStateAfterTurn1 = enactor.enactTurn(turn1, initialMatchState);
+    var matchStateAfterTurn2 = enactor.enactTurn(turn2, matchStateAfterTurn1);
+    var matchStateAfterTurn3 = enactor.enactTurn(turn3, matchStateAfterTurn2);
 //    var matchStateAfterTurn4 = matchStateAfterTurn3.doTurn(turn4);
 
     var matchFormatter = new MatchFormatter();
